Improve formatter (#145)
* Don’t run formatter directly on the saved file

The previous code formatter was triggered after a file was saved to the
disk. After that, the `mix format` command was run on that file,
formatting its content and changing them on the disk. At last, VSCode
reloaded the file with the updated contents.

This approach had two major disadvantages:
- The whole process is slow and has a delay of about 2 seconds on every
  file save.
- Since the changes to the file happen outside of the editor, we break
  the undo feature and cause some trouble with other extensions such as
  live-share.

With this commit we use stdio to pass and get the code from `mix format`
instead of formatting directly the file on the disk. The resulting
formatted code is managed by VSCode, allowing undos and increasing
performance.