Updates package.json postinstall to prepare
The npm `postinstall` script runs both on local npm install as well as
after being installed as a dependency of someone else’s project. The
`prepare` script only runs on local npm install (and before publishing).
[npm-scripts](https://docs.npmjs.com/cli/v6/using-npm/scripts)

`vscode` is a `devDependency`, not a `dependency`, so if someone tries to
install this package as a dependency of their own, the installation
fails because `./node_modules/vscode` does not exist.

One example of where it may be useful to install the package directly
would be when using something like
[andrewbranch/gatsby-remark-vscode](https://github.com/andrewbranch/gatsby-remark-vscode)
which uses VS Code syntax highlighting and themes to generate the syntax
highlighting in Gatsby sites.

The `postinstall` to `prepare` change is
[suggested](https://github.com/andrewbranch/gatsby-remark-vscode/issues/129)
by the author of the Gatsby plugin as a way to work with languates that
aren't supported by VS Code out-of-the-box.

This change seems to have works successfully in the past for other VS
Code extensions as well:

* https://github.com/juanfranblanco/vscode-solidity/pull/186
* https://github.com/styled-components/vscode-styled-components/pull/203